requirements.txt
blinker==1.6.2
cachelib==0.10.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.4
Flask-Session==0.4.1
grpcio==1.54.0
grpcio-tools==1.54.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
protobuf==4.22.3
requests==2.28.2
urllib3==1.26.15
waitress==2.1.2
Werkzeug==2.3.2
zipp==3.15.0
